requirements.txt
fastapi
uvicorn
whisper
faster-whisper
pydantic
python-dotenv
elasticsearch
ffmpeg-python
pytesseract
opencv-python
reportlab
fpdf2
requests
skimage.metrics
python-docx